delete old validator class
