+++ tests/test_examples.py
@@ -5,7 +5,7 @@
 from collections import OrderedDict
 
 from validatedata import validate, validate_data
-from validatedata.validator import ValidationError
+from validatedata import ValidationError
 from .base import BaseTest
