+++ validatedata/__init__.py
@@ -1,5 +1,5 @@
 from .validatedata import validate, validate_data, validate_types, ValidationResult, VALID_RULE_KEYS, check_rule
-from .validator import ValidationError
+from .engine import ValidationError
 
 __version__ = '0.4.1'
